backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
ultralytics==8.2.0
insightface==0.7.3
onnxruntime==1.18.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pandas==2.2.2
